+++ paddlevideo/loader/dataset/msrvtt.py
@@ -19,13 +19,15 @@ import numpy as np
 try:
     import lmdb
 except ImportError as e:
-    print(f"{e}, [lmdb] package and it's dependencies is required for ActBERT.")
+    print(
+        f"Warning! {e}, [lmdb] package and it's dependencies is required for ActBERT."
+    )
 import pickle
 try:
     from paddlenlp.transformers import BertTokenizer
 except ImportError as e:
     print(
-        f"{e}, [paddlenlp] package and it's dependencies is required for ActBERT."
+        f"Warning! {e}, [paddlenlp] package and it's dependencies is required for ActBERT."
     )
 from ..registry import DATASETS
 from .base import BaseDataset
