+++ paddlevideo/loader/pipelines/decode.py
@@ -17,7 +17,7 @@ try:
     import av
 except ImportError as e:
     print(
-        f"{e}, [av] package and it's dependencies is required for TimeSformer and other models."
+        f"Warning! {e}, [av] package and it's dependencies is required for TimeSformer and other models."
     )
 import cv2
 import pickle
