+++ setup.py
@@ -17,6 +17,7 @@ from io import open
 with open('requirements.txt', encoding="utf-8-sig") as f:
     requirements = f.readlines()
 
+
 def readme():
     with open('docs/en/whl_en.md', encoding="utf-8-sig") as f:
         README = f.read()
@@ -24,13 +25,14 @@ def readme():
 
 
 setup(
-    name='paddlevideo', #name of .whl file
-    packages=['ppvideo'], #install package name
+    name='paddlevideo',  #name of .whl file
+    packages=['ppvideo'],  #install package name
     package_dir={'ppvideo': ''},
-    include_package_data=True, #Accept all data files and directories matched by MANIFEST.in
+    include_package_data=
+    True,  #Accept all data files and directories matched by MANIFEST.in
     install_requires=requirements,
-    entry_points={"console_scripts": ["ppvideo= ppvideo.tools.paddlevideo_clas:main"]},
-    version='0.0.1',
+    entry_points={"console_scripts": ["ppvideo= ppvideo.tools.wheel:main"]},
+    version='1.0.0',
     license='Apache License 2.0',
     description='Awesome Video toolkits based on PaddlePaddle ',
     long_description=readme(),
@@ -38,7 +40,7 @@ setup(
     url='https://github.com/PaddlePaddle/PaddleVideo',
     download_url='https://github.com/PaddlePaddle/PaddleVideo.git',
     keywords=[
-    'A treasure chest for video understanding powered by PaddlePaddle.'
+        'A treasure chest for video understanding powered by PaddlePaddle.'
     ],
     classifiers=[
         'Intended Audience :: Developers', 'Operating System :: OS Independent',
@@ -50,4 +52,5 @@ setup(
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7', 'Topic :: Utilities'
-    ],)
+    ],
+)
