+++ tools/utils.py
@@ -23,14 +23,14 @@ try:
     import imageio
 except ImportError as e:
     print(
-        f"{e}, [imageio] package and it's dependencies is required for VideoSwin."
+        f"Warning! {e}, [imageio] package and it's dependencies is required for VideoSwin."
     )
 try:
     import matplotlib as mpl
     import matplotlib.cm as cm
 except ImportError as e:
     print(
-        f"{e}, [matplotlib] package and it's dependencies is required for ADDS."
+        f"Warning! {e}, [matplotlib] package and it's dependencies is required for ADDS."
     )
 import numpy as np
 import paddle
@@ -54,70 +54,11 @@ from paddlevideo.metrics.bmn_metric import boundary_choose, soft_nms
 from paddlevideo.utils import Registry, build, get_config
 from paddlevideo.modeling.framework.segmenters.utils import ASRFPostProcessing
 
-from ava_predict import (detection_inference, frame_extraction,
-                         get_detection_result, get_timestep_result, pack_result,
-                         visualize)
+from .ava_predict import (detection_inference, frame_extraction,
+                          get_detection_result, get_timestep_result,
+                          pack_result, visualize)
 
 INFERENCE = Registry('inference')
-
-
-def decode(filepath, args):
-    num_seg = args.num_seg
-    seg_len = args.seg_len
-
-    cap = cv2.VideoCapture(filepath)
-    videolen = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-    sampledFrames = []
-    for i in range(videolen):
-        ret, frame = cap.read()
-        # maybe first frame is empty
-        if ret == False:
-            continue
-        img = frame[:, :, ::-1]
-        sampledFrames.append(img)
-    average_dur = int(len(sampledFrames) / num_seg)
-    imgs = []
-    for i in range(num_seg):
-        idx = 0
-        if average_dur >= seg_len:
-            idx = (average_dur - 1) // 2
-            idx += i * average_dur
-        elif average_dur >= 1:
-            idx += i * average_dur
-        else:
-            idx = i
-
-        for jj in range(idx, idx + seg_len):
-            imgbuf = sampledFrames[int(jj % len(sampledFrames))]
-            img = Image.fromarray(imgbuf, mode='RGB')
-            imgs.append(img)
-
-    return imgs
-
-
-def preprocess(img, args):
-    img = {"imgs": img}
-    resize_op = Scale(short_size=args.short_size)
-    img = resize_op(img)
-    ccrop_op = CenterCrop(target_size=args.target_size)
-    img = ccrop_op(img)
-    to_array = Image2Array()
-    img = to_array(img)
-    if args.normalize:
-        img_mean = [0.485, 0.456, 0.406]
-        img_std = [0.229, 0.224, 0.225]
-        normalize_op = Normalization(mean=img_mean, std=img_std)
-        img = normalize_op(img)
-    return img['imgs']
-
-
-def postprocess(output, args):
-    output = output.flatten()
-    output = F.softmax(paddle.to_tensor(output)).numpy()
-    classes = np.argpartition(output, -args.top_k)[-args.top_k:]
-    classes = classes[np.argsort(-output[classes])]
-    scores = output[classes]
-    return classes, scores
 
 
 def build_inference_helper(cfg):
@@ -177,7 +118,8 @@ class Base_Inference_helper():
 
     def postprocess(self,
                     output: np.ndarray,
-                    print_output: bool = True) -> None:
+                    print_output: bool = True,
+                    return_result: bool = False):
         """postprocess
 
         Args:
@@ -195,15 +137,25 @@ class Base_Inference_helper():
                                     list(output.shape[1:]))  # [N, T, C]
             output = output.mean(axis=1)  # [N, C]
         output = F.softmax(paddle.to_tensor(output), axis=-1).numpy()
+        results_list = []
         for i in range(N):
             classes = np.argpartition(output[i], -self.top_k)[-self.top_k:]
             classes = classes[np.argsort(-output[i, classes])]
             scores = output[i, classes]
+            topk_class = classes[:self.top_k]
+            topk_scores = scores[:self.top_k]
+            result = {
+                "video_id": self.input_file[i],
+                "topk_class": topk_class,
+                "topk_scores": topk_scores
+            }
+            results_list.append(result)
             if print_output:
                 print("Current video file: {0}".format(self.input_file[i]))
-                for j in range(self.top_k):
-                    print("\ttop-{0} class: {1}".format(j + 1, classes[j]))
-                    print("\ttop-{0} score: {1}".format(j + 1, scores[j]))
+                print("\ttop-{0} class: {1}".format(self.top_k, topk_class))
+                print("\ttop-{0} score: {1}".format(self.top_k, topk_scores))
+        if return_result:
+            return results_list
 
 
 @INFERENCE.register()
@@ -231,7 +183,7 @@ class ppTSM_Inference_helper(Base_Inference_helper):
         img_mean = [0.485, 0.456, 0.406]
         img_std = [0.229, 0.224, 0.225]
         ops = [
-            VideoDecoder(),
+            VideoDecoder("decord"),
             Sampler(self.num_seg, self.seg_len, valid_mode=True),
             Scale(self.short_size),
             CenterCrop(self.target_size),
@@ -375,9 +327,9 @@ class BMN_Inference_helper(Base_Inference_helper):
 
         json.dump(result_dict, outfile)
 
+
 @INFERENCE.register()
 class TokenShift_Inference_helper(Base_Inference_helper):
-
     def __init__(self,
                  num_seg=8,
                  seg_len=1,
@@ -404,9 +356,7 @@ class TokenShift_Inference_helper(Base_Inference_helper):
         results = {'filename': input_file}
         ops = [
             VideoDecoder(backend='pyav', mode='test', num_seg=self.num_seg),
-            Sampler(self.num_seg,
-                    self.seg_len,
-                    valid_mode=True),
+            Sampler(self.num_seg, self.seg_len, valid_mode=True),
             Normalization(self.mean, self.std, tensor_shape=[1, 1, 1, 3]),
             Image2Array(data_format='cthw'),
             JitterScale(self.short_size, self.short_size),
@@ -836,7 +786,6 @@ class AGCN2s_Inference_helper(Base_Inference_helper):
             input_file)
         data = np.load(input_file)
         results = {'data': data}
-        
 
         res = np.expand_dims(results['data'], axis=0).copy()
         return [res]
@@ -1093,7 +1042,7 @@ class TransNetV2_Inference_helper():
             import ffmpeg
         except ImportError as e:
             print(
-                f"{e}, [ffmpeg-python] package and it's dependencies is required for TransNetV2."
+                f"Warning! {e}, [ffmpeg-python] package and it's dependencies is required for TransNetV2."
             )
         assert os.path.isfile(input_file) is not None, "{0} not exists".format(
             input_file)
